requirements.txt
Flask==3.1.0
groq
python-dotenv==1.0.1
gunicorn==23.0.0
pypdf==5.1.0
python-docx==1.1.2